requirements.txt
streamlit==1.51.0
torch
accelerate
safetensors
diffusers>=0.35.0
transformers>=4.57.0
huggingface-hub
Pillow
fpdf
open-clip-torch
pydantic>=2.7
